requirements/prod.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile --find-links=https://download.pytorch.org/whl/torch_stable.html requirements/prod.in
#
--find-links https://download.pytorch.org/whl/torch_stable.html

absl-py==1.0.0
    # via
    #   tensorboard
    #   tensorflow
aiohttp==3.8.1
    # via fsspec
aiosignal==1.2.0
    # via aiohttp
astunparse==1.6.3
    # via tensorflow
async-timeout==4.0.2
    # via aiohttp
attrs==21.4.0
    # via aiohttp
bcrypt==3.2.2
    # via paramiko
beautifulsoup4==4.11.1
    # via gdown
boltons==21.0.0
    # via -r requirements/prod.in
cachetools==5.0.0
    # via google-auth
certifi==2021.10.8
    # via requests
cffi==1.15.0
    # via
    #   bcrypt
    #   cryptography
    #   pynacl
charset-normalizer==2.0.12
    # via
    #   aiohttp
    #   requests
click==8.1.3
    # via flask
confluent-kafka==1.8.2
    # via -r requirements/prod.in
cryptography==37.0.1
    # via paramiko
deepface==0.0.74
    # via -r requirements/prod.in
editdistance==0.6.0
    # via -r requirements/prod.in
filelock==3.6.0
    # via gdown
flask==2.1.2
    # via
    #   -r requirements/prod.in
    #   deepface
flatbuffers==2.0
    # via tensorflow
frozenlist==1.3.0
    # via
    #   aiohttp
    #   aiosignal
fsspec[http]==2022.3.0
    # via pytorch-lightning
gast==0.5.3
    # via tensorflow
gdown==4.4.0
    # via
    #   deepface
    #   retina-face
google-auth==2.6.6
    # via
    #   google-auth-oauthlib
    #   tensorboard
google-auth-oauthlib==0.4.6
    # via tensorboard
google-pasta==0.2.0
    # via tensorflow
grpcio==1.44.0
    # via
    #   tensorboard
    #   tensorflow
h5py==3.6.0
    # via
    #   -r requirements/prod.in
    #   tensorflow
idna==3.3
    # via
    #   requests
    #   yarl
importlib-metadata==4.11.3
    # via
    #   flask
    #   markdown
imutils==0.5.4
    # via -r requirements/prod.in
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
keras==2.8.0
    # via
    #   deepface
    #   mtcnn
    #   tensorflow
keras-preprocessing==1.1.2
    # via tensorflow
libclang==14.0.1
    # via tensorflow
markdown==3.3.6
    # via tensorboard
markupsafe==2.1.1
    # via jinja2
mtcnn==0.1.1
    # via deepface
multidict==6.0.2
    # via
    #   aiohttp
    #   yarl
numpy==1.21.6
    # via
    #   -r requirements/prod.in
    #   deepface
    #   h5py
    #   keras-preprocessing
    #   opencv-python
    #   opt-einsum
    #   pandas
    #   pytorch-lightning
    #   retina-face
    #   tensorboard
    #   tensorflow
    #   torch
    #   torchmetrics
    #   torchvision
oauthlib==3.2.0
    # via requests-oauthlib
opencv-python==4.5.5.64
    # via
    #   -r requirements/prod.in
    #   deepface
    #   mtcnn
    #   retina-face
opt-einsum==3.3.0
    # via tensorflow
packaging==21.3
    # via
    #   pytorch-lightning
    #   torchmetrics
pandas==1.3.5
    # via deepface
paramiko==2.10.4
    # via -r requirements/prod.in
patool==1.12
    # via -r requirements/prod.in
pillow==9.1.0
    # via
    #   deepface
    #   retina-face
    #   torchvision
protobuf==3.20.1
    # via
    #   tensorboard
    #   tensorflow
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pycparser==2.21
    # via cffi
pydeprecate==0.3.2
    # via
    #   pytorch-lightning
    #   torchmetrics
pynacl==1.5.0
    # via paramiko
pyparsing==3.0.8
    # via packaging
pysocks==1.7.1
    # via requests
python-dateutil==2.8.2
    # via pandas
python-dotenv==0.20.0
    # via -r requirements/prod.in
pytorch-lightning==1.6.2
    # via -r requirements/prod.in
pytz==2022.1
    # via pandas
pyyaml==6.0
    # via pytorch-lightning
rarfile==4.0
    # via -r requirements/prod.in
requests[socks]==2.27.1
    # via
    #   -r requirements/prod.in
    #   fsspec
    #   gdown
    #   requests-oauthlib
    #   tensorboard
requests-oauthlib==1.3.1
    # via google-auth-oauthlib
retina-face==0.0.10
    # via deepface
rsa==4.8
    # via google-auth
six==1.16.0
    # via
    #   absl-py
    #   astunparse
    #   gdown
    #   google-auth
    #   google-pasta
    #   grpcio
    #   keras-preprocessing
    #   paramiko
    #   python-dateutil
    #   tensorflow
smart-open==6.0.0
    # via -r requirements/prod.in
soupsieve==2.3.2.post1
    # via beautifulsoup4
tensorboard==2.8.0
    # via
    #   pytorch-lightning
    #   tensorflow
tensorboard-data-server==0.6.1
    # via tensorboard
tensorboard-plugin-wit==1.8.1
    # via tensorboard
tensorflow==2.8.0
    # via
    #   deepface
    #   retina-face
tensorflow-io-gcs-filesystem==0.25.0
    # via tensorflow
termcolor==1.1.0
    # via tensorflow
tf-estimator-nightly==2.8.0.dev2021122109
    # via tensorflow
toml==0.10.2
    # via -r requirements/prod.in
torch==1.8.1+cu101
    # via
    #   -r requirements/prod.in
    #   pytorch-lightning
    #   torchaudio
    #   torchmetrics
    #   torchvision
torch-metrics==1.1.7
    # via -r requirements/prod.in
torchaudio==0.8.1
    # via -r requirements/prod.in
torchmetrics==0.8.1
    # via pytorch-lightning
torchvision==0.9.1+cu101
    # via -r requirements/prod.in
tqdm==4.64.0
    # via
    #   deepface
    #   gdown
    #   pytorch-lightning
typing-extensions==4.2.0
    # via
    #   pytorch-lightning
    #   tensorflow
    #   torch
urllib3==1.26.9
    # via requests
werkzeug==2.1.2
    # via
    #   flask
    #   tensorboard
wget==3.2
    # via -r requirements/prod.in
wheel==0.37.1
    # via
    #   astunparse
    #   tensorboard
wrapt==1.14.1
    # via tensorflow
yarl==1.7.2
    # via aiohttp
zipp==3.8.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
